+++ CHANGELOG.md
@@ -2041,3 +2041,4 @@
 - [2026-06-16 17:13] perf(ui): apply RepaintBoundary on animated ui widgets
 - [2026-06-16 19:03] fix(navigation): correct layout constraints and overflow in navigation
 - [2026-06-16 19:10] fix(core): correct layout constraints and overflow in core
+- [2026-06-16 19:30] feat(i18n): implement responsive i18n view component